Update penalty error and clean up
This makes the penalty error larger so that on the weird degree 7 polynomial we came up with in class there isn't a tendency to just "take the error" because that's a nice big number that's close to the target values.

I also removed some commented out examples of evolved programs that aren't relevant to Fizz-Buzz.