Create our first successful hyper inteligent AI
The major thing we did was change the `string_removechar`
to `string_removesubstring` because working with chars
was annoying to debug. The new function just replaces
particular characters with a blank.

We then removed some things like boolsm 0/1, exec_if and
did some runs, but we forgot to write down how many
generations those took.